# openapi.yaml
openapi: 3.1.0
info:
  title: Panda API
  version: 1.0.0
  description: |
    Public JSON API for Panda desktop and integrations. Configure clients with
    an `apiUrl` pointing to `/api/v1` and an API token in the Bearer header.
    Protocol design heavily references https://github.com/tianma-if/edgeever.
servers:
  - url: http://127.0.0.1:8787/api/v1
security:
  - bearerAuth: []
paths:
  /mcp:
    post:
      summary: Stateless MCP Streamable HTTP JSON-RPC endpoint
      responses:
        '200': { description: JSON-RPC response }
        '204': { description: Notification accepted without a response body }
        '401': { description: Bearer token required }
  /api/v1/mcp:
    post:
      summary: Alias of the stateless MCP Streamable HTTP endpoint
      responses:
        '200': { description: JSON-RPC response }
        '204': { description: Notification accepted without a response body }
        '401': { description: Bearer token required }
...
  /memos/{id}/save:
    post:
      summary: Save memo using ETag/revision optimistic concurrency
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: Save acknowledgement }, '409': { description: Revision conflict } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: API token }